+++ src/aosm/azext_aosm/_params.py
@@ -145,3 +145,41 @@ def load_arguments(self: AzCommandsLoader, _):
                 "Requires Docker to be installed locally."
             ),
         )
+
+    with self.argument_context("aosm sns") as c:
+        c.argument(
+            "output_file",
+            options_list=["--output-file"],
+            help="The name of the output file to write the generated config text to.",
+            required=False,
+        )
+        c.argument(
+            "config_file",
+            options_list=["--config-file", "-f"],
+            type=file_type,
+            completer=FilesCompleter(allowednames="*.jsonc"),
+            help=(
+                "The path to the configuration file. This is a JSONC file that contains"
+                " the required parameters for building the NSD."
+            ),
+        )
+        c.argument(
+            "build_output_folder",
+            options_list=["--build-output-folder", "-b"],
+            type=file_type,
+            completer=FilesCompleter(allowednames="*.json"),
+            help="Path to the folder to deploy, created by the build command.",
+        )
+        c.argument(
+            "no_subscription_permissions",
+            options_list=["--no-subscription-permissions", "-u"],
+            completer=FilesCompleter(allowednames="*.json"),
+            help=(
+                "Pass this flag if you do not have permission to import to the "
+                "Publisher subscription (Contributor role + AcrPush role, or a "
+                "custom role that allows the importImage action and AcrPush over the "
+                "whole subscription). This means that the image artifacts will be "
+                "pulled to your local machine and then pushed to the Artifact Store. "
+                "Requires Docker to be installed locally."
+            ),
+        )