+++ src/aosm/HISTORY.rst
@@ -2,7 +2,11 @@
 
 Release History
 ===============
-
+2.0.0b2
+++++++++
+* added sns generate-config command
+* added a check to make sure resource type used in ARM template are in allowed list
+* Fixed multi NF RETs issue in nsdvs
 2.0.0b1
 ++++++++
 * Renamed nfdvName to nfdv in CGVs
@@ -35,6 +39,10 @@ Release History
 * Added validation of the values file for helm charts when using the `az aosm nfd build` command for the `cnf` definition type
 * Fixed helm chart image parsing in the `az aosm nfd build` command for the `cnf` definition type. This means that the images can now be extracted correctly from the helm chart.
 * Fixed: infinite loop bug when retrying failed artifact uploads to the ACR
+* Added a check to determine if all base resources are already deployed. If they all are, the deployment of base bicep is skipped.
+* Added: Build and Deploy operations for site resource.
+    * `az aosm site build` to prepare files for a site.
+    * `az aosm site deploy` to deploy a site.
 
 1.0.0b4
 ++++++++